pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "addressforge"
version = "0.1.0"
description = "Open-source self-hosted address intelligence platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic",
  "python-dotenv",
  "mysql-connector-python",
  "requests",
  "numpy",
  "pandas",
  "scipy",
  "postal",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
